pyproject.toml
[tool.poetry]
name = "school"
version = "0.1.0"
description = ""
authors = ["faiver90 <faiver90@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
django = ">=5.2.8,<6.0.0"
ruff = ">=0.14.5,<0.15.0"
mypy = ">=1.18.2,<2.0.0"
gunicorn = ">=23.0.0,<24.0.0"
uvicorn = ">=0.38.0,<0.39.0"
psycopg = {extras = ["binary"], version = "^3.2.12"}
djangorestframework = ">=3.16.1,<4.0.0"
celery = ">=5.5.3,<6.0.0"
redis = ">=7.0.1,<8.0.0"
flower = ">=2.0.1,<3.0.0"
drf-spectacular = ">=0.29.0,<0.30.0"
drf-spectacular-sidecar = ">=2025.10.1,<2026.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
